requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
h5py==3.11.0
idna==3.10
numpy==1.24.4
pillow==10.4.0
PyYAML==6.0.2
requests==2.32.3
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchvision==0.13.1+cu102
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
